+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Add sale order and client partner fields on production order.
 - Add generated purchase orders panel in sale order form.
 - COST SHEET : calculation date
 - INVOICE : filter on product in InvoiceLine on supplier catalog for supplier invoice
+- Message : Email - 'From' e-mail adress contact Name.
 
 ## Improvements
 - FINANCIAL ACCOUNT : New boolean 'Analytic Distribution Authorized' added to make Analytic fields required on Invoice Line.